rabbitmq_policy support version 3.8.0
In RabbitMQ version 3.8.0 rabbitmqctl output is not in json anymore.
So rabbitmqctl status returns plain text and current search return
None instead of rabbit version. This results in getting wrong
policies inside _list_policies as we need to omit first line from 3.7.9

While PR #62996 is not merged, this PR worth placing, and
is pretty backportable.

Signed-off-by: Dmitriy Rabotyagov <noonedeadpunk@ya.ru>